added folders and moved methods
